chore: refactoring the code
